requirements.txt
autopep8==1.4.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-dotenv==0.10.3
SQLAlchemy==1.3.11
Werkzeug==0.16.0
